+++ src/CodeProject.ObjectPool/PooledObject.cs
@@ -48,11 +48,31 @@ namespace CodeProject.ObjectPool
         #region Internal Methods - resource and state management
 
         /// <summary>
-        ///   Validates pooled object state.
+        ///   Validates pooled object state. An invalid object will not get into the pool and it will
+        ///   not be returned to consumers.
         /// </summary>
         /// <param name="validationContext">The validation context.</param>
         /// <returns>True if current pooled object is valid, false otherwise.</returns>
-        protected internal virtual bool ValidateObject(PooledObjectValidationContext validationContext) => true;
+        internal bool ValidateObject(PooledObjectValidationContext validationContext)
+        {
+            if (OnValidateObject != null)
+            {
+                try
+                {
+                    return OnValidateObject(validationContext);
+                }
+                catch (Exception ex)
+                {
+#if !NET35
+                    if (Log.IsWarnEnabled()) Log.WarnException("[ObjectPool] An unexpected error occurred while validating an object", ex);
+#else
+                    System.Diagnostics.Debug.Assert(ex != null); // Placeholder to avoid warnings
+#endif
+                    return false;
+                }
+            }
+            return true;
+        }
 
         /// <summary>
         ///   Releases the object resources. This method will be called by the pool manager when
@@ -70,10 +90,7 @@ namespace CodeProject.ObjectPool
                 catch (Exception ex)
                 {
 #if !NET35
-                    if (Log.IsWarnEnabled())
-                    {
-                        Log.WarnException("[ObjectPool] An unexpected error occurred while releasing resources", ex);
-                    }
+                    if (Log.IsWarnEnabled()) Log.WarnException("[ObjectPool] An unexpected error occurred while releasing resources", ex);
 #else
                     System.Diagnostics.Debug.Assert(ex != null); // Placeholder to avoid warnings
 #endif
@@ -99,25 +116,10 @@ namespace CodeProject.ObjectPool
                 {
                     OnResetState();
                 }
-                catch (CannotResetStateException crsex)
-                {
-#if !NET35
-                    if (Log.IsDebugEnabled())
-                    {
-                        Log.DebugException("[ObjectPool] Object state could not be reset", crsex);
-                    }
-#else
-                    System.Diagnostics.Debug.Assert(crsex != null); // Placeholder to avoid warnings
-#endif
-                    return false;
-                }
                 catch (Exception ex)
                 {
 #if !NET35
-                    if (Log.IsWarnEnabled())
-                    {
-                        Log.WarnException("[ObjectPool] An unexpected error occurred while resetting state", ex);
-                    }
+                    if (Log.IsWarnEnabled()) Log.WarnException("[ObjectPool] An unexpected error occurred while resetting state", ex);
 #else
                     System.Diagnostics.Debug.Assert(ex != null); // Placeholder to avoid warnings
 #endif
@@ -130,6 +132,12 @@ namespace CodeProject.ObjectPool
         #endregion Internal Methods - resource and state management
 
         #region Events - extending resource and state management
+
+        /// <summary>
+        ///   Validates pooled object state. An invalid object will not get into the pool and it will
+        ///   not be returned to consumers.
+        /// </summary>
+        public Func<PooledObjectValidationContext, bool> OnValidateObject { get; set; }
 
         /// <summary>
         ///   Reset the object state to allow this object to be re-used by other parts of the application.
@@ -171,10 +179,7 @@ namespace CodeProject.ObjectPool
             catch (Exception ex)
             {
 #if !NET35
-                if (Log.IsWarnEnabled())
-                {
-                    Log.WarnException("[ObjectPool] An error occurred while re-adding to pool", ex);
-                }
+                if (Log.IsWarnEnabled()) Log.WarnException("[ObjectPool] An error occurred while re-adding to pool", ex);
 #else
                 System.Diagnostics.Debug.Assert(ex != null); // Placeholder to avoid warnings
 #endif